Make UI only draw in Viewport Toolbar.
Per Blender's standards the UI should not create duplicate panels in multiple regions, this is confusing for the user and unnecessarily inflates the UI.

Since the Asset Manager is used primarily for adding objects to the scene it belongs in the Toolbar. Whether if should exist in the *Tools* tab or the *Create* tab is debatable.

Ref #3